♻️ refactor: move errorResponse
