+++ src/app/(backend)/api/middleware/auth/index.ts
@@ -2,10 +2,10 @@ import { AuthObject } from '@clerk/backend';
 import { getAuth } from '@clerk/nextjs/server';
 import { NextRequest } from 'next/server';
 
-import { createErrorResponse } from '@/app/(backend)/api/errorResponse';
 import { JWTPayload, LOBE_CHAT_AUTH_HEADER, OAUTH_AUTHORIZED, enableClerk } from '@/const/auth';
 import { AgentRuntime, AgentRuntimeError, ChatCompletionErrorPayload } from '@/libs/agent-runtime';
 import { ChatErrorType } from '@/types/fetch';
+import { createErrorResponse } from '@/utils/errorResponse';
 
 import { checkAuthMethod, getJWTPayload } from './utils';
